backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
google-genai
Pillow
python-multipart
httpx
